movie overview and user are correctly rerendered on state change
