+++ src/components/Dashboard/MovieOnDashboard/RemoveMovieFromDashboard/RemoveMovieFromDashboard.tsx
@@ -1,10 +1,8 @@
-import React, {useContext} from 'react';
+import React from 'react';
 import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
 import styled from 'styled-components';
 import {useSnackbar} from 'notistack';
 import {IMovie} from '../../../../movieseat';
-import localforage from 'localforage';
-import {MovieContext} from '../../../../context/MovieContext';
 import {useMutation} from '@apollo/client';
 import resolvers from '../../../../../src/resolvers';
 import {movieVar} from '../../../../cache';
@@ -21,23 +19,11 @@ const DeleteButton = styled.div`
 `;
 const RemoveMovieFromDashboard = ({movie}: {movie: IMovie}) => {
   const [removeMovieRes] = useMutation(resolvers.mutations.RemoveMovie);
-  const [, setMovies] = useContext(MovieContext);
   const {enqueueSnackbar} = useSnackbar();
 
   const removeMovieFromList = async (movie) => {
     removeMovieRes({variables: {id: parseInt(movie.id)}});
     movieVar(movie);
-    const value = await localforage.getItem<IMovie []>('trackedMovies');
-    if (value) {
-      value.forEach((item, index) => {
-        if (item.id === movie.id) {
-          value.splice(index, 1);
-          localforage.setItem('trackedMovies', value);
-          setMovies(value);
-        }
-      });
-    }
-
     enqueueSnackbar(
         'Removed ' + movie.original_title + ' from your watchlist.',
         {variant: 'success'},
