+++ src/components/Header/Header.tsx
@@ -2,8 +2,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import {Link} from 'react-router-dom';
-import {useReactiveVar} from '@apollo/client';
-import {currentUserVar} from '../../cache';
 import Login from '../Login/Login';
 import {Route} from 'react-router-dom';
 
@@ -20,19 +18,10 @@ const Title = styled.h1`
 `;
 
 const Header = () => {
-  const currentUser = useReactiveVar(currentUserVar);
-  const logout = (event) => {
-    event.preventDefault();
-    window.localStorage.removeItem('token');
-    window.localStorage.removeItem('user_id');
-    currentUserVar({id: 0, email: '', isLoggedIn: false});
-  };
-
   return (
     <HeaderStyle>
-      <Route path="/login" exact component={Login} />
+      <Route path="/" exact component={Login} />
       <Link to="/"><Title>Movieseat</Title></Link>
-      { currentUser.isLoggedIn ? <a href="" onClick={logout}>Logout</a> : <Link to="/login">login</Link> }
     </HeaderStyle>
   );
 };
